Back-django/requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.3.0
factory-boy==2.12.0
Pillow==6.2.0
